+++ resources/compiledJS/localisation/codex_localisation.js
@@ -1,6 +1,20 @@
 "use strict";
 const codexLang = {
-    finnish: {},
-    english: {}
+    finnish: {
+        ITEMS_title: "ESINEET",
+        ENEMIES_title: "VIHOLLISET",
+        SUMMONS_title: "KUTSUTTAVAT",
+        CLASSES_title: "LUOKAT",
+        ABILITIES_title: "KYVYT",
+        CHARACTERS_title: "HAHMOT",
+    },
+    english: {
+        ITEMS_title: "ITEMS",
+        ENEMIES_title: "ENEMIES",
+        SUMMONS_title: "SUMMONS",
+        CLASSES_title: "CLASSES",
+        ABILITIES_title: "ABILITIES",
+        CHARACTERS_title: "CHARACTERS",
+    }
 };
 //# sourceMappingURL=codex_localisation.js.map