+++ script/localisation/codex_localisation.ts
@@ -1,4 +1,23 @@
+interface codexLoc {
+  [finnish: string]: any;
+  english: any;
+}
+
 const codexLang = {
-  finnish: {},
-  english: {}
-};
+  finnish: {
+    ITEMS_title: "ESINEET",
+    ENEMIES_title: "VIHOLLISET",
+    SUMMONS_title: "KUTSUTTAVAT",
+    CLASSES_title: "LUOKAT",
+    ABILITIES_title: "KYVYT",
+    CHARACTERS_title: "HAHMOT",
+  },
+  english: {
+    ITEMS_title: "ITEMS",
+    ENEMIES_title: "ENEMIES",
+    SUMMONS_title: "SUMMONS",
+    CLASSES_title: "CLASSES",
+    ABILITIES_title: "ABILITIES",
+    CHARACTERS_title: "CHARACTERS",
+  }
+} as codexLoc;