+++ script/ui/save.ts
@@ -116,6 +116,7 @@ async function gotoSaveMenu(inMainMenu = false, animate: boolean = true) {
         loadingScreen.style.display = "none";
         return warningMessage("<i>resources/icons/error.png<i>Failed to load save.\nIt may be corrupted or too old.");
       }
+      closeSaveMenu();
       player = pl;
       fallenEnemies = fe;
       itemData = id;