+++ Changelog.md
@@ -14,6 +14,7 @@ The new changes can be gamebreaking.
 - Added `Central Heere` map.
 - Added `entrances` which are gateways to adjacent maps.
 - Added new character `Thrisna`. She runs a store in the village of Myre.
+- Expanded codex with new information.
 
 ## Fixes & Tweaks
 
@@ -33,6 +34,7 @@ The new changes can be gamebreaking.
 - Fixed cooldowns resetting when loading saved game.
 - `Passive abilities` are now called `Traits`.
 - Female characters no longer render nude when wearing only pants.
+- Added a lot of missing localisation
 
 # 1.0.9 ~ 16.3.2022